Move status / statusError into internal status package
